src/scheduler/requirements.txt
nats-py==2.10.0
python-dotenv==1.1.1
APScheduler==3.11.0
protobuf==6.31.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
qdrant-client==1.14.3
sentence-transformers==5.0.0
playwright==1.53.0